Move components and hooks into individual files
