+++ src/renderer/sidebar/src/components/AssistantMessage.tsx
@@ -1,0 +1,19 @@
+import React from 'react'
+import { StreamingText } from './StreamingText'
+import { Markdown } from './Markdown'
+
+// Assistant Message Component - appears on the left
+export const AssistantMessage: React.FC<{ content: string; isStreaming?: boolean }> = ({
+    content,
+    isStreaming
+}) => (
+    <div className="relative w-full animate-fade-in">
+        <div className="py-1">
+            {isStreaming ? (
+                <StreamingText content={content} />
+            ) : (
+                <Markdown content={content} />
+            )}
+        </div>
+    </div>
+)
