+++ src/renderer/sidebar/src/components/LoadingIndicator.tsx
@@ -1,0 +1,20 @@
+import React, { useState, useEffect } from 'react'
+import { cn } from '@common/lib/utils'
+
+// Loading Indicator with spinning star
+export const LoadingIndicator: React.FC = () => {
+    const [isVisible, setIsVisible] = useState(false)
+
+    useEffect(() => {
+        setIsVisible(true)
+    }, [])
+
+    return (
+        <div className={cn(
+            "transition-transform duration-300 ease-in-out",
+            isVisible ? "scale-100" : "scale-0"
+        )}>
+            ...
+        </div>
+    )
+}
