+++ src/renderer/sidebar/src/components/Markdown.tsx
@@ -1,0 +1,49 @@
+import React from 'react'
+import ReactMarkdown from 'react-markdown'
+import remarkGfm from 'remark-gfm'
+import remarkBreaks from 'remark-breaks'
+
+// Markdown Renderer Component
+export const Markdown: React.FC<{ content: string }> = ({ content }) => (
+    <div className="prose prose-sm dark:prose-invert max-w-none 
+                    prose-headings:text-foreground prose-p:text-foreground 
+                    prose-strong:text-foreground prose-ul:text-foreground 
+                    prose-ol:text-foreground prose-li:text-foreground
+                    prose-a:text-primary hover:prose-a:underline
+                    prose-code:bg-muted prose-code:px-1 prose-code:py-0.5 
+                    prose-code:rounded prose-code:text-sm prose-code:text-foreground
+                    prose-pre:bg-muted dark:prose-pre:bg-muted/50 prose-pre:p-3 
+                    prose-pre:rounded-lg prose-pre:overflow-x-auto">
+        <ReactMarkdown
+            remarkPlugins={[remarkGfm, remarkBreaks]}
+            components={{
+                // Custom code block styling
+                code: ({ node, className, children, ...props }) => {
+                    const inline = !className
+                    return inline ? (
+                        <code className="bg-muted dark:bg-muted/50 px-1 py-0.5 rounded text-sm text-foreground" {...props}>
+                            {children}
+                        </code>
+                    ) : (
+                        <code className={className} {...props}>
+                            {children}
+                        </code>
+                    )
+                },
+                // Custom link styling
+                a: ({ children, href }) => (
+                    <a
+                        href={href}
+                        className="text-primary hover:underline"
+                        target="_blank"
+                        rel="noopener noreferrer"
+                    >
+                        {children}
+                    </a>
+                ),
+            }}
+        >
+            {content}
+        </ReactMarkdown>
+    </div>
+)
