+++ src/renderer/sidebar/src/components/StreamingText.tsx
@@ -1,0 +1,26 @@
+import React, { useState, useEffect } from 'react'
+
+// Streaming Text Component
+export const StreamingText: React.FC<{ content: string }> = ({ content }) => {
+    const [displayedContent, setDisplayedContent] = useState('')
+    const [currentIndex, setCurrentIndex] = useState(0)
+
+    useEffect(() => {
+        if (currentIndex < content.length) {
+            const timer = setTimeout(() => {
+                setDisplayedContent(content.slice(0, currentIndex + 1))
+                setCurrentIndex(currentIndex + 1)
+            }, 10)
+            return () => clearTimeout(timer)
+        }
+    }, [content, currentIndex])
+
+    return (
+        <div className="whitespace-pre-wrap text-foreground">
+            {displayedContent}
+            {currentIndex < content.length && (
+                <span className="inline-block w-2 h-5 bg-primary/60 dark:bg-primary/40 ml-0.5 animate-pulse" />
+            )}
+        </div>
+    )
+}
