+++ src/renderer/sidebar/src/components/types.ts
@@ -1,0 +1,7 @@
+export interface Message {
+    id: string
+    role: 'user' | 'assistant'
+    content: string
+    timestamp: number
+    isStreaming?: boolean
+}
