+++ src/renderer/sidebar/src/components/useAutoScroll.ts
@@ -1,0 +1,22 @@
+import { useRef, useLayoutEffect } from 'react'
+import { Message } from './types'
+
+// Auto-scroll hook
+export const useAutoScroll = (messages: Message[]) => {
+    const scrollRef = useRef<HTMLDivElement>(null)
+    const prevCount = useRef(0)
+
+    useLayoutEffect(() => {
+        if (messages.length > prevCount.current) {
+            setTimeout(() => {
+                scrollRef.current?.scrollIntoView({
+                    behavior: 'smooth',
+                    block: 'end'
+                })
+            }, 100)
+        }
+        prevCount.current = messages.length
+    }, [messages.length])
+
+    return scrollRef
+}
